+++ src/layouts/InterfaceLayout/containers/SwapContainer/SwapContainer.vue
@@ -995,6 +995,12 @@ export default {
           break;
       }
     },
+    decimalsGenerally(toCurrency) {
+      if(toCurrency.toLowerCase() === 'neo'){
+        return 0;
+      }
+      return this.fiatCurrenciesArray.includes(toCurrency) ? 2 : 6;
+    },
     async updateRateEstimate(fromCurrency, toCurrency, fromValue, to) {
       if (this.haveProviderRates) {
         this.showWarning = this.warningCurrencies.includes(toCurrency);
@@ -1048,11 +1054,7 @@ export default {
                   minValue: entry.minValue || 0,
                   maxValue: entry.maxValue || 0,
                   computeConversion: _fromValue => {
-                    const decimals = this.fiatCurrenciesArray.includes(
-                      toCurrency
-                    )
-                      ? 2
-                      : 6;
+                    const decimals = this.decimalsGenerally(toCurrency)
                     return new BigNumber(_fromValue)
                       .times(entry.rate)
                       .toFixed(decimals)